+++ Services/CMdm.Services/CustomModule/Fcmb/AccountOfficerService.cs
@@ -88,7 +88,7 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">recordId identifier</param>
         /// <returns>Vendor</returns>
-        public virtual IList<AccountOfficer> GetAccountOfficerbyIds(int[] recordIds)
+        public virtual IList<AccountOfficer> GetAccountOfficerbyIds(Int64[] recordIds)
         {
             if (recordIds == null || recordIds.Length == 0)
                 return null;
@@ -100,7 +100,7 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">MdmDQQue identifier</param>
         /// <returns>Vendor</returns>
-        public virtual AccountOfficer GetAccOffbyId(int recordId)
+        public virtual AccountOfficer GetAccOffbyId(Int64 recordId)
         {
             if (recordId == 0)
                 return null;
