+++ Services/CMdm.Services/CustomModule/Fcmb/IWrongSchemeCodeService.cs
@@ -16,13 +16,13 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">recordId identifier</param>
         /// <returns>Vendor</returns>
-        IList<WrongSchemeCode> GetWrongSchemeCodebyIds(int[] recordIds);
+        IList<WrongSchemeCode> GetWrongSchemeCodebyIds(Int64[] recordIds);
         /// <summary>
         /// Gets a Queitem by item reference identifier
         /// </summary>
         /// <param name="recordId">que identifier</param>
         /// <returns>Vendor</returns>
-        WrongSchemeCode GetWrongSchCodebyId(int recordId);
+        WrongSchemeCode GetWrongSchCodebyId(Int64 recordId);
         /// <summary>
         /// Gets all items
         /// </summary>
